Catch if a Spark job really fails; resolve #76. (#80)

+++ app/jobs/collections_spark_job.rb
@@ -41,14 +41,19 @@ class CollectionsSparkJob < ApplicationJob
       spark_job_cmd = spark_shell + ' --master local[' + spark_threads + '] --driver-memory ' + spark_memory_driver + ' --conf spark.network.timeout=' + spark_network_timeout + ' --packages "io.archivesunleashed:aut:' + aut_version + '" -i ' + collection_spark_job_file + ' | tee ' + collection_spark_job_file + '.log'
       logger.info 'Executing: ' + spark_job_cmd
       system(spark_job_cmd)
-      combine_full_url_output_cmd = 'cat ' + collection_derivatives + '/all-domains/output/part* > ' + collection_derivatives + '/all-domains/' + c.collection_id.to_s + '-fullurls.txt'
-      logger.info 'Executing: ' + combine_full_url_output_cmd
-      system(combine_full_url_output_cmd)
-      FileUtils.rm_rf(collection_derivatives + '/all-domains/output')
-      combine_full_text_output_cmd = 'cat ' + collection_derivatives + '/all-text/output/part* > ' + collection_derivatives + '/all-text/' + c.collection_id.to_s + '-fulltext.txt'
-      logger.info 'Executing: ' + combine_full_text_output_cmd
-      system(combine_full_text_output_cmd)
-      FileUtils.rm_rf(collection_derivatives + '/all-text/output')
+      successful_job = collection_derivatives + '/all-domains/output/_SUCCESS'
+      if File.exist? successful_job
+        combine_full_url_output_cmd = 'cat ' + collection_derivatives + '/all-domains/output/part* > ' + collection_derivatives + '/all-domains/' + c.collection_id.to_s + '-fullurls.txt'
+        logger.info 'Executing: ' + combine_full_url_output_cmd
+        system(combine_full_url_output_cmd)
+        FileUtils.rm_rf(collection_derivatives + '/all-domains/output')
+        combine_full_text_output_cmd = 'cat ' + collection_derivatives + '/all-text/output/part* > ' + collection_derivatives + '/all-text/' + c.collection_id.to_s + '-fulltext.txt'
+        logger.info 'Executing: ' + combine_full_text_output_cmd
+        system(combine_full_text_output_cmd)
+        FileUtils.rm_rf(collection_derivatives + '/all-text/output')
+      else
+        raise 'Collections spark job failed.'
+      end
     end
   end
 end
